removed addr parsing from lib
